Renamed filters module to signal and moved related functions.
