+++ qats/ts.py
@@ -12,8 +12,8 @@ from scipy.interpolate import interp1d
 from scipy.signal import welch
 from scipy.stats import kurtosis, skew, tstd
 from .rainflow import count_cycles
-from .filters import lowpass, highpass, bandblock, bandpass, threshold as thresholdpass, smooth, taper
-from .stats import average_frequency, find_maxima
+from .signal import lowpass, highpass, bandblock, bandpass, threshold as thresholdpass, smooth, taper, \
+    average_frequency, find_maxima
 from .weibull import Weibull, weibull2gumbel, pwm
 from .gumbel import Gumbel
